Remove a sentence
> The Django community has earned a reputation of being welcoming to
beginners and we would like to keep it that way.

This may be true but the CoC is more about protecting individuals from
sustaining condecendance from others than it is about the community's
reputation. It may be very dangerous to imply the community reputation
is one of our objectives, because we may do very wrong things to protect
it.